+++ crates/converter/src/formats/mod.rs
@@ -1,7 +1,10 @@
+pub(crate) mod cbor;
 pub(crate) mod json;
 pub(crate) mod message_pack;
 pub(crate) mod xml;
 
+#[cfg(test)]
+mod cbor_tests;
 #[cfg(test)]
 mod json_tests;
 #[cfg(test)]