+++ crates/converter/src/shared/mod.rs
@@ -1,2 +1,4 @@
+pub(crate) mod binary_gate_type;
 pub(crate) mod circuit_data;
 pub(crate) mod gate_operation_data;
+pub(crate) mod readable_gate_type;